fixed compile issues with Linux GCC compiler

+++ include/coredev/timer.h
@@ -208,7 +208,7 @@ public:
 	virtual uint32_t EnableTimerTrigger(int TrigNo, uint32_t msPeriod, TIMER_TRIG_TYPE Type,
 	                           	   	    TIMER_TRIGCB const Handler = NULL, void * const pContext = NULL)
 	{
-		return (uint32_t)(EnableTimerTrigger(TrigNo, (uint64_t)msPeriod * 1000000ULL, Type, Handler, pContext) / 1000000ULL);
+		return (uint32_t)((uint64_t )EnableTimerTrigger(TrigNo, (uint64_t)((uint64_t)msPeriod * 1000000ULL), Type, Handler, pContext) / 1000000ULL);
 	}
 
 	/**
